VideoCommon: Add helpers for generating common render states
